+++ editor/build-non-split.js
@@ -13,6 +13,9 @@ config.optimization.runtimeChunk = false;
 
 // JS
 config.output.filename = 'static/js/main.js';
-// CSS. "5" is MiniCssPlugin
 
+// CSS. "5" is MiniCssPlugin
 config.plugins[5].options.filename = 'static/css/main.css';
+
+// Fix the file size embedding limit
+config.module.rules[2].oneOf[0].options.limit = 50000;